ai_interview_backend/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
celery
redis
django-redis
python-dotenv
openai
pypdf
python-docx
djangorestframework-simplejwt
dj-rest-auth
django-allauth
drf-nested-routers